Move the setOption util to the common mixin
This work moves the private setOption "util" to it's only use as a private method on the common mixin.  It also moved the `mergeOptions` into the `setOptions` util as it was always called immediately afterwards and the naming still makes sense.
