--- src/utils/set-options.js
@@ -1,8 +1,0 @@
-import _ from 'underscore';
-
-// Internal utility for setting options consistently across Mn
-const setOptions = function(options) {
-  this.options = _.extend({}, _.result(this, 'options'), options);
-};
-
-export default setOptions;
